Added missing virtual destructors. Modified VertexData and IndexData copy constructors so that they'll throw a warning at link time if used. Fixed a scoping issue in GpuProgramManager. Fixed a unsigned/signed comparision warning.

+++ OgreMain/src/OgreMesh.cpp
@@ -958,7 +958,7 @@ namespace Ogre {
             unsigned char* pDest = static_cast<unsigned char*>(
                 newBuffer->lock(HardwareBuffer::HBL_DISCARD));
             size_t vertSize = origBuffer->getVertexSize();
-            for (int v = 0; v < vertexData->vertexCount; ++v)
+            for (size_t v = 0; v < vertexData->vertexCount; ++v)
             {
                 // Copy original vertex data
                 memcpy(pDest, pSrc, vertSize);
